Rename the component and break it into sub-components
